✨ feat(core): migrate to PrintingFfi instance for better management
- Refactor the example app to use the singleton `PrintingFfi.instance`
- Update method calls to use `PrintingFfi.instance`
- Update the example app to use new models
- Add exception models
- Add printer properties result model
