+++ lib/models/models.dart
@@ -1,0 +1,7 @@
+export 'printer.dart';
+export 'print_job.dart';
+export 'print_options.dart';
+export 'pdf_print_settings.dart';
+export 'printer_capabilities.dart';
+export 'exceptions.dart';
+export 'printer_properties_result.dart';
